chore: refactor component to use CSR
