+++ app/leaderboard/Leaderboard.tsx
@@ -1,5 +1,5 @@
-import { Leaderboards } from "../_utils/global";
-import { getLeaderboardData } from "../_utils/fetchLeaderboard";
+"use client";
+import React, { useEffect, useState } from "react";
 import {
   Table,
   TableHeader,
@@ -9,15 +9,43 @@ import {
   TableCell,
 } from "../_components/ui/table";
 import FooterMenu from "../_components/FooterMenu";
+import { Leaderboards } from "../_utils/global";
+import { AuthContext } from "../_components/useContext/AuthContext";
 
-export default async function Leaderboard() {
-  const LeaderboardData: Leaderboards[] | undefined =
-    await getLeaderboardData();
+interface LeaderboardProps {}
 
-  if (!LeaderboardData) {
-    alert("Leaderboard currently unavailble");
-    return;
-  }
+const Leaderboard: React.FC<LeaderboardProps> = () => {
+  const storedName = localStorage.getItem("username");
+  console.log(storedName);
+
+  const [leaderboardData, setLeaderboardData] = useState<Leaderboards[]>([]);
+
+  const getLeaderboardData = async () => {
+    try {
+      const baseUrl = process.env.NEXT_PUBLIC_BASE_URL;
+      const cacheBuster = Date.now();
+      const response = await fetch(
+        `${baseUrl}/api/leaderboards?_=${cacheBuster}`,
+        {
+          credentials: "include",
+          method: "GET",
+          headers: {
+            "Content-Type": "application/json",
+            "Cache-Control": "no-cache",
+          },
+        }
+      );
+      const resData: Leaderboards[] = await response.json() as Leaderboards [];
+      setLeaderboardData(resData);
+    } catch (error) {
+      alert("Leaderboard currently unavailable");
+      console.error("Error fetching leaderboard:", error);
+    }
+  };
+
+  useEffect(() => {
+    void getLeaderboardData();
+  }, []);
 
   return (
     <>
@@ -35,10 +63,18 @@ export default async function Leaderboard() {
             </TableRow>
           </TableHeader>
           <TableBody>
-            {LeaderboardData.map((player, index) => (
+            {leaderboardData.map((player, index) => (
               <TableRow key={index}>
                 <TableCell>{index + 1}</TableCell>
-                <TableCell>{player.username}</TableCell>
+                <TableCell
+                  className={
+                    player.username === storedName
+                      ? "text-blue-600 font-bold bg-yellow-100" // Example: Applying background color to highlight
+                      : ""
+                  }
+                >
+                  {player.username}
+                </TableCell>
                 <TableCell>{player.score}</TableCell>
               </TableRow>
             ))}
@@ -50,4 +86,6 @@ export default async function Leaderboard() {
       <FooterMenu variant="leaderboard" />
     </>
   );
-}
+};
+
+export default Leaderboard;
